Unit Info Update + BalanceSheetCalculator (#492)
* Update the BalanceSheetCalculator & Units' node info

update ConsumerUnit

Update StorageUnit

update VehicleUnit

Update DistributionUnit

update imports format

update ProductUnit

Update unit_info from dict to dataclasses

Update BalnanceSheetCalculator

Update EnvSampler & EnvHelper & ...

Update hello example

fix storage initialization bug; update supply chain test

update notebook

update import format

uadd tick parameter to balance calculator

update the ProductUnit.get_sale_mean()

remove duplicated profit and cost part from facility

* Fix typing & formatting issues in PR 492 (#497)

* Update the BalanceSheetCalculator & Units' node info

update ConsumerUnit

Update StorageUnit

update VehicleUnit

Update DistributionUnit

update imports format

update ProductUnit

Update unit_info from dict to dataclasses

Update BalnanceSheetCalculator

Update EnvSampler & EnvHelper & ...

Update hello example

fix storage initialization bug; update supply chain test

update notebook

update import format

uadd tick parameter to balance calculator

update the ProductUnit.get_sale_mean()

remove duplicated profit and cost part from facility

* Code polishment

* A minor comment resolved

+++ maro/simulator/scenarios/supply_chain/actions.py
@@ -1,5 +1,6 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
+
 from dataclasses import dataclass
 
 
@@ -18,4 +19,4 @@ class ConsumerAction(SupplyChainAction):
 
 @dataclass
 class ManufactureAction(SupplyChainAction):
-    production_rate: float
+    production_rate: int
